Apply Code review suggestions
